requirements.txt
dask
distributed
snowflake-connector-python[pandas]
snowflake-sqlalchemy